feat: add close function to anchor_vault_q3_2024
This commit adds a new function called `close` to the `anchor_vault_q3_2024` module. The `close` function allows users to close their vault and transfer the remaining funds to their account. This function is implemented in the `Close` struct and uses the CPI (Cross-Program Invocation) mechanism to transfer the funds from the vault to the user's account.